bpf: tests: collapse legacy DSR-GENEVE test
Have a single file for the IPv4 and IPv6 tests.
